refac(structure): Keep DEDSTA structures in separate module. Add quick_test to gitignore.
